# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PortFolioGen AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type GeneratePortfolioInputTemplate = typeof GeneratePortfolioInputTemplate[keyof typeof GeneratePortfolioInputTemplate];


export const GeneratePortfolioInputTemplate = {
  dark: 'dark',
  minimal: 'minimal',
  creative: 'creative',
} as const;

export interface GeneratePortfolioInput {
  /** @minLength 1 */
  prompt: string;
  template?: GeneratePortfolioInputTemplate;
}

export interface GeneratedPortfolio {
  html: string;
  css: string;
  js: string;
}

export interface Project {
  id: number;
  /** @nullable */
  userId?: string | null;
  title: string;
  prompt: string;
  html: string;
  css: string;
  js: string;
  /** @nullable */
  template?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  prompt: string;
  html: string;
  css: string;
  js: string;
  template?: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  title?: string;
  html?: string;
  css?: string;
  js?: string;
}

export interface ProjectStats {
  total: number;
  recentCount: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};