# lib/api-zod/src/generated/types/generatePortfolioInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PortFolioGen AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratePortfolioInputTemplate } from './generatePortfolioInputTemplate';

export interface GeneratePortfolioInput {
  /** @minLength 1 */
  prompt: string;
  template?: GeneratePortfolioInputTemplate;
}