# lib/api-zod/src/generated/types/mobileTokenExchangeSuccess.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PortFolioGen AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MobileTokenExchangeSuccess {
  token: string;
}