# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PortFolioGen AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Project {
  id: number;
  /** @nullable */
  userId?: string | null;
  title: string;
  prompt: string;
  html: string;
  css: string;
  js: string;
  /** @nullable */
  template?: string | null;
  createdAt: Date;
  updatedAt: Date;
}